+++ examples/debug.R
@@ -1,4 +1,3 @@
-detach("package:myClim", unload=TRUE)
 setwd("tests/testthat")
 devtools::load_all()