+++ tests/testthat/test-prep.R
@@ -157,6 +157,7 @@ test_that("mc_prep_crop", {
     expect_equal(length(cropped_data$localities$A2E32$loggers[[1]]$datetime), 8)
     expect_equal(length(cropped_data$localities$A2E32$loggers[[1]]$sensors$TMS_T1$values), 8)
     expect_equal(cropped_data$localities$A2E32$loggers[[1]]$sensors$TMS_T1$states$end, lubridate::ymd_h("2020-10-16 08"))
+    expect_equal(length(cropped_data$localities$A1E05$loggers$Thermo_1$datetime), 0)
     cropped_data <- mc_prep_crop(data, end=as.POSIXct("2020-10-16 08:00", tz="UTC"), end_included=FALSE)
     test_raw_data_format(cropped_data)
     expect_equal(length(cropped_data$localities$A2E32$loggers[[1]]$datetime), 7)
@@ -195,26 +196,10 @@ test_that("mc_prep_crop by localities", {
     expect_error(cropped_data <- mc_prep_crop(data,
                                               start=as.POSIXct(c("2020-10-28 09:00", "2020-10-16 07:00"), tz="UTC"),
                                               end=as.POSIXct(c("2020-10-28 10:00", "2020-10-16 08:00"), tz="UTC")))
-    cropped_data <- mc_prep_crop(data,
-                                 start=as.POSIXct(c("2020-10-28 09:00", "2020-10-16 07:00"), tz="UTC"),
-                                 end=as.POSIXct(c("2020-10-28 10:00", "2020-10-16 08:00"), tz="UTC"),
-                                 localities=c("A1E05", "A2E32"))
-    test_raw_data_format(cropped_data)
-    expect_equal(cropped_data$localities$A1E05$loggers[[1]]$datetime,
-                 as.POSIXct(c("2020-10-28 09:00", "2020-10-28 09:15", "2020-10-28 09:30", "2020-10-28 09:45", "2020-10-28 10:00"), tz="UTC"))
-    expect_equal(cropped_data$localities$A2E32$loggers[[1]]$datetime,
-                 as.POSIXct(c("2020-10-16 07:00", "2020-10-16 07:15", "2020-10-16 07:30", "2020-10-16 07:45", "2020-10-16 08:00"), tz="UTC"))
-    expect_equal(cropped_data$localities$A6W79$loggers[[1]]$datetime, data$localities$A6W79$loggers[[1]]$datetime)
-    cropped_agg_data <- mc_prep_crop(agg_data,
-                                     start=as.POSIXct(c("2020-10-28 09:00", "2020-10-16 07:00"), tz="UTC"),
-                                     end=as.POSIXct(c("2020-10-28 10:00", "2020-10-16 08:00"), tz="UTC"),
-                                     localities=c("A1E05", "A2E32"))
-    test_agg_data_format(cropped_agg_data)
-    expect_equal(cropped_agg_data$localities$A1E05$datetime,
-                 as.POSIXct(c("2020-10-28 09:00", "2020-10-28 09:15", "2020-10-28 09:30", "2020-10-28 09:45", "2020-10-28 10:00"), tz="UTC"))
-    expect_equal(cropped_agg_data$localities$A2E32$datetime,
-                 as.POSIXct(c("2020-10-16 07:00", "2020-10-16 07:15", "2020-10-16 07:30", "2020-10-16 07:45", "2020-10-16 08:00"), tz="UTC"))
-    expect_equal(cropped_agg_data$localities$A6W79$datetime, agg_data$localities$A6W79$datetime)
+    expect_error(cropped_data <- mc_prep_crop(data,
+                                              start=as.POSIXct(c("2020-10-28 09:00", "2020-10-16 07:00"), tz="UTC"),
+                                              end=as.POSIXct(c("2020-10-28 10:00", "2020-10-16 08:00"), tz="UTC"),
+                                              localities=c("A1E05", "A2E32")))
 })
 
 test_that("mc_prep_crop errors", {
@@ -230,6 +215,40 @@ test_that("mc_prep_crop errors", {
                                lubridate::ymd_hm("2022-02-24 09:15"),
                                lubridate::ymd_hm("2022-02-24 09:45"),
                                lubridate::ymd_hm("2022-02-24 10:30")))
+})
+
+test_that("mc_prep_crop crop_table", {
+    data <- mc_read_data("../data/TOMST/files_table2.csv", clean=TRUE, silent=TRUE)
+    crop_table <- as.data.frame(tibble::tribble(
+        ~locality_id,  ~logger_name, ~start, ~end,
+        "A1E05"     ,       "TMS_1", lubridate::ymd_h("2022-04-07 08"), lubridate::ymd_h("2022-04-07 10"),
+        "A2E32"     , NA_character_, lubridate::ymd_h("2020-10-16 08"),                                NA,
+        "A6W79"     ,   "TMS_L45_1",                                NA, lubridate::ymd_h("2020-10-16 10"),
+    ))
+    cropped_data <- mc_prep_crop(data, crop_table=crop_table)
+    test_raw_data_format(cropped_data)
+    loggers <- mc_info_logger(cropped_data)
+    expect_equal(loggers$start_date, c(lubridate::ymd_hm("2020-10-28 08:45"),
+                                       lubridate::ymd_h("2022-04-07 08"),
+                                       lubridate::ymd_h("2020-10-16 08"),
+                                       lubridate::ymd_h("2020-10-16 00")))
+    expect_equal(loggers$end_date, c(lubridate::ymd_hm("2020-10-28 11:15"),
+                                     lubridate::ymd_h("2022-04-07 10"),
+                                     lubridate::ymd_hm("2020-10-17 00:45"),
+                                     lubridate::ymd_h("2020-10-16 10")))
+    crop_table <- as.data.frame(tibble::tribble(
+        ~locality_id,  ~logger_name, ~start, ~end,
+        "A1E05"     ,       "TMS_1", lubridate::ymd_h("2022-04-07 08"), lubridate::ymd_h("2022-04-07 10"),
+        "A1E05"     ,       "TMS_1", lubridate::ymd_h("2022-04-07 09"), lubridate::ymd_h("2022-04-07 10"),
+    ))
+    expect_error(cropped_data <- mc_prep_crop(data, crop_table=crop_table))
+})
+
+test_that("mc_prep_crop crop_table 2", {
+    data <- mc_load("../data/crop/data_cs_1.rds")
+    crop_table <- readRDS("../data/crop/crop_table_cs_1.rds")
+    crop_data <- mc_prep_crop(data, crop_table=crop_table)
+    test_raw_data_format(crop_data)
 })
 
 test_that(".prep_get_loggers_datetime_step_unprocessed", {